Now compiling on OS X/Mac
Finally got the program compiling on OS X/Mac. Using a path to the static version of c++experimental for now: "/usr/local/opt/llvm/lib/libc++experimental.a". Requires the newest version of clang/LLVM which is not provided by xcode. Version used for compilation is v6.0.1, and was installed via homebrew using the command: brew install llvm